Update 06_TrakingSkimageLib.py
In the first part,

(H, hogImage) = feature.hog(gray, orientations=9, pixels_per_cell=(8, 8),    cells_per_block=(2, 2),
   visualize=True)

we define the main parameters of the model, and the number of orientations will be 9 as we explained before. We will have 64 pixels per cell and now we can divide the blocks into cells of 16 16, so we define blocks as 2 by 2; each one of 88 pixels.